Move fixture to common module
